requirements.txt
fastapi>=0.95.0,<0.100
uvicorn[standard]>=0.22.0,<0.23
SQLAlchemy>=1.4,<2.0
python-jose[cryptography]>=3.0.0
passlib[bcrypt]>=1.7.0,<1.8
bcrypt>=4.0.0,<5.0
python-multipart
pydantic>=2.0.0,<3.0
email-validator
redis>=4.5.0
# testing
pytest>=7.4.0
httpx>=0.24.0,<0.28
